Mental_health_detection_using_face_expression/requirements.txt
deepface==0.0.81
fastapi==0.109.0
mongoengine==0.27.0
numpy==1.26.3
opencv_python==4.9.0.80
passlib==1.7.4
Pillow==10.2.0
pydantic==2.6.0
PyJWT==2.8.0
python_jose==3.3.0
